pyproject.toml
[tool.poetry]
name = "hello-fastapi-poetry"
version = "0.1.0"
description = "A simple FastAPI app with Docker and Poetry"
authors = ["varshakumarr <varshaprv19@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.116.0"}
uvicorn = "^0.35.0"
pydantic = "^2.11.7"
pydantic-settings = "^2.10.1"
sqlalchemy = "^2.0.41"
asyncpg = "^0.30.0"
greenlet = "^3.2.3"
python-dotenv = "^1.1.1"
wikipedia = "^1.4.0"
pyyaml = "^6.0.2"
ollama = "^0.5.1"
pgvector = "^0.2.5"
langchain = "^0.2.0"
langchain-text-splitters = "^0.2.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"